Vagrantfile to provide portable development environment
The Vagrantfile uses the script vagrant.sh to provision a Debian Linux
system to run PyExZ3. It could be tweaked to configure just about any
Linux system.

The .vagrant folder has been added to .gitignore as well.

Additionally I added `tmp.txt` to the ignore file as it is the output of
one of the test cases. Perhaps the tests would be better run inside of a
sandbox.