Added CVC String Backslash Handling
CVC's string API uses the same parser in the string constructor as the
SMT-LIB standard. This can cause execution mismatches and PyEx
crashes when the string does not comply with SMT-LIB.

For now I've added a single check for backslashes, but there are a bunch
more cases that are not handled correctly. See
http://cvc4.cs.nyu.edu/wiki/Strings#Escape_Sequences_for_String_Literals
. Additionally there's an assert on the length of the string matching
the reported size by CVC, which should catch some errors in the future.